Make scope work with nil values
